Factor out type inference for expressions
